+++ db/projects/project-teams/project-members/e_project_member.sql
@@ -5,7 +5,7 @@ CREATE TABLE projects.e_project_member (
   id BIGSERIAL,
   e_project_team_id BIGINT NOT NULL,
   e_person_id BIGINT NOT NULL,
-  d_project_member_state_id INTEGER NOT NULL,
+  d_project_member_state_id BIGINT NOT NULL,
     is_deleted BOOLEAN NOT NULL DEFAULT FALSE,
       PRIMARY KEY (
         id
